+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `RestMgmtInterface` endpoint
 - BREAKING CHANGE: Removed fixed setting `AAAPolicy` from `InteropService`
 - BREAKING CHANGE: Removed fixed setting `ACL` from `SSHService`
+- BREAKING CHANGE: Removed fixed setting `ACL` from `WebGUI`
 
 ## v0.15.0